+++ quick/app/Http/Controllers/ClinicController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Clinic;
 use App\Campaign;
-
+use Auth;
 
 class ClinicController extends Controller
 {
@@ -42,30 +42,35 @@ class ClinicController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => ['required', 'string', 'max:100'],
-            'clinic_record' => ['required', 'string', 'max:100'],
-            'technical_manager' => ['required', 'string', 'max:100'],
-            'professional_record' => ['required', 'string', 'max:100']
-        ],[
-            'name.required' => 'O campo Nome é obrigatório.',
-            'name.min' => 'Desculpe, mas o nome deve possuir no mínimo 10 caracteres.',
-            'name.max' => 'Desculpe o nome excede o limite de caracteres.',
-
-            'clinic_record.required' => 'O campo Registro clínico é obrigatório.',
-            'clinic_record.max' => 'Desculpe o Registro clínico excede o limite de caracteres.',
-
-            'technical_manager.required' => 'O campo Responsável técnico é obrigatório.',
-            'technical_manager.max' => 'Desculpe o campo Responsável técnico excede o limite de caracteres.',
-
-            'professional_record.required' => 'O campo Registro profissional é obrigatório.',
-            'professional_record.max' => 'Desculpe o campo Registro profissional excede o limite de caracteres.'
-
-        ]);
-
-        Clinic::create($request->all());
-
-        return back()->with('success','Cadastro efetuado com sucesso!');
+        if(Auth::user()->type_user == 0) {
+            $request->validate([
+                'name' => ['required', 'string', 'max:100'],
+                'clinic_record' => ['required', 'string', 'max:100'],
+                'technical_manager' => ['required', 'string', 'max:100'],
+                'professional_record' => ['required', 'string', 'max:100']
+            ],[
+                'name.required' => 'O campo Nome é obrigatório.',
+                'name.min' => 'Desculpe, mas o nome deve possuir no mínimo 10 caracteres.',
+                'name.max' => 'Desculpe o nome excede o limite de caracteres.',
+    
+                'clinic_record.required' => 'O campo Registro clínico é obrigatório.',
+                'clinic_record.max' => 'Desculpe o Registro clínico excede o limite de caracteres.',
+    
+                'technical_manager.required' => 'O campo Responsável técnico é obrigatório.',
+                'technical_manager.max' => 'Desculpe o campo Responsável técnico excede o limite de caracteres.',
+    
+                'professional_record.required' => 'O campo Registro profissional é obrigatório.',
+                'professional_record.max' => 'Desculpe o campo Registro profissional excede o limite de caracteres.'
+    
+            ]);
+    
+            Clinic::create($request->all());
+    
+            return back()->with('success','Cadastro efetuado com sucesso!');
+        } else {
+            return back()->with('error','Usuário não autorizado.');
+        }
+       
     }
 
     /**
@@ -99,33 +104,37 @@ class ClinicController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => ['required', 'string', 'max:100'],
-            'clinic_record' => ['required', 'string', 'max:100'],
-            'technical_manager' => ['required', 'string', 'max:100'],
-            'professional_record' => ['required', 'string', 'max:100']
-        ],[
-            'name.required' => 'O campo Nome é obrigatório.',
-            'name.min' => 'Desculpe, mas o nome deve possuir no mínimo 10 caracteres.',
-            'name.max' => 'Desculpe o nome excede o limite de caracteres.',
-
-            'clinic_record.required' => 'O campo Registro clínico é obrigatório.',
-            'clinic_record.min' => 'Desculpe, mas o campo Registro clínico deve possuir no mínimo 10 caracteres.',
-            'clinic_record.max' => 'Desculpe o Registro clínico excede o limite de caracteres.',
-
-            'technical_manager.required' => 'O campo Responsável técnico é obrigatório.',
-            'technical_manager.min' => 'Desculpe, mas o campo Responsável técnico deve possuir no mínimo 10 caracteres.',
-            'technical_manager.max' => 'Desculpe o campo Responsável técnico excede o limite de caracteres.',
-
-            'professional_record.required' => 'O campo Registro profissional é obrigatório.',
-            'professional_record.min' => 'Desculpe, mas o campo Registro profissional deve possuir no mínimo 10 caracteres.',
-            'professional_record.max' => 'Desculpe o campo Registro profissional excede o limite de caracteres.'
-        ]);
-
-        $clinic = Clinic::find($id); 
-        $clinic->update($request->all());
-
-        return back()->with('success','Post atualizado com sucesso!');
+        if(Auth::user()->type_user == 0) {
+            $request->validate([
+                'name' => ['required', 'string', 'max:100'],
+                'clinic_record' => ['required', 'string', 'max:100'],
+                'technical_manager' => ['required', 'string', 'max:100'],
+                'professional_record' => ['required', 'string', 'max:100']
+            ],[
+                'name.required' => 'O campo Nome é obrigatório.',
+                'name.min' => 'Desculpe, mas o nome deve possuir no mínimo 10 caracteres.',
+                'name.max' => 'Desculpe o nome excede o limite de caracteres.',
+    
+                'clinic_record.required' => 'O campo Registro clínico é obrigatório.',
+                'clinic_record.min' => 'Desculpe, mas o campo Registro clínico deve possuir no mínimo 10 caracteres.',
+                'clinic_record.max' => 'Desculpe o Registro clínico excede o limite de caracteres.',
+    
+                'technical_manager.required' => 'O campo Responsável técnico é obrigatório.',
+                'technical_manager.min' => 'Desculpe, mas o campo Responsável técnico deve possuir no mínimo 10 caracteres.',
+                'technical_manager.max' => 'Desculpe o campo Responsável técnico excede o limite de caracteres.',
+    
+                'professional_record.required' => 'O campo Registro profissional é obrigatório.',
+                'professional_record.min' => 'Desculpe, mas o campo Registro profissional deve possuir no mínimo 10 caracteres.',
+                'professional_record.max' => 'Desculpe o campo Registro profissional excede o limite de caracteres.'
+            ]);
+    
+            $clinic = Clinic::find($id); 
+            $clinic->update($request->all());
+    
+            return back()->with('success','Post atualizado com sucesso!');
+        } else {
+            return back()->with('error','Usuário não autorizado.');
+        }        
     }
 
     /**
@@ -136,9 +145,13 @@ class ClinicController extends Controller
      */
     public function destroy($id)
     {
-        $clinic = Clinic::find($id); 
-        $clinic->delete();
+        if(Auth::user()->type_user == 0) {
+            $clinic = Clinic::find($id); 
+            $clinic->delete();
 
-        return redirect()->back()->with('success','Clínica excluída com sucesso!');
+            return redirect()->back()->with('success','Clínica excluída com sucesso!');
+        } else {
+            return back()->with('error','Usuário não autorizado.');
+        }     
     }
 }
