+++ quick/app/Http/Controllers/PostController.php
@@ -10,6 +10,7 @@ use App\CampaignPost;
 use App\Figure;
 use Storage;
 use App\Clinic;
+use Auth;
 
 class PostController extends Controller
 {
@@ -47,39 +48,43 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'color' => ['required', 'string', 'max:100'],
-            'campaign' => ['required', 'string', 'max:100'],
-            'logo' => ['required', 'string', 'max:10'],
-            'image' => ['required', 'image', 'mimes:jpeg,jpg,png,gif,svg|max:5000|dimensions:min_width=100,min_height=100,max_width=5000,max_height=5000']
-
-        ],[
-            'color.required' => 'O campo Cor é obrigatório.',
-            'campaign.required' => 'O campo Campanha é obrigatório.',
-            'image.required' => 'O campo Imagem é obrigatório.',
-            'image.mimes' => 'Somente arquivos de imagem são aceitos: jpeg, jpg, png, gif, svg.',
-            'image.max' => 'Somente são aceitas images de até 5MB.'
-        ]);
-
-        // Verifica se informou o arquivo e se é válido
-        if ($request->hasFile('image')) 
-        {
-           
-            if($request->image->isValid())
-            {
-                $post = Post::create(
-                    ['color' => $request->color, 'logo' => $request->logo, 'campaign_id' => $request->campaign, 'image' => $request->image->store('images')]
-                );
-            }
-                unset($image);
-        }
-
-        CampaignPost::create(
-            ['campaign_id' => $request->campaign, 'post_id' => $post->id]
-        );
+        if(Auth::user()->type_user == 0) {
+            $request->validate([
+                'color' => ['required', 'string', 'max:100'],
+                'campaign' => ['required', 'string', 'max:100'],
+                'logo' => ['required', 'string', 'max:10'],
+                'image' => ['required', 'image', 'mimes:jpeg,jpg,png,gif,svg|max:5000|dimensions:min_width=100,min_height=100,max_width=5000,max_height=5000']
     
-
-        return back()->with('success','Cadastro efetuado com sucesso!');
+            ],[
+                'color.required' => 'O campo Cor é obrigatório.',
+                'campaign.required' => 'O campo Campanha é obrigatório.',
+                'image.required' => 'O campo Imagem é obrigatório.',
+                'image.mimes' => 'Somente arquivos de imagem são aceitos: jpeg, jpg, png, gif, svg.',
+                'image.max' => 'Somente são aceitas images de até 5MB.'
+            ]);
+    
+            // Verifica se informou o arquivo e se é válido
+            if ($request->hasFile('image')) 
+            {
+               
+                if($request->image->isValid())
+                {
+                    $post = Post::create(
+                        ['color' => $request->color, 'logo' => $request->logo, 'campaign_id' => $request->campaign, 'image' => $request->image->store('images')]
+                    );
+                }
+                    unset($image);
+            }
+    
+            CampaignPost::create(
+                ['campaign_id' => $request->campaign, 'post_id' => $post->id]
+            );
+        
+    
+            return back()->with('success','Cadastro efetuado com sucesso!');
+        } else {
+            return back()->with('error','Usuário não autorizado.');
+        }
     }
 
     /**
@@ -109,38 +114,42 @@ class PostController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => ['string', 'max:100'],
-            'color' => ['string', 'max:100'],
-        ],[
-            'name.max' => 'Desculpe o nome excede o limite de caracteres.'
-        ]);
-        
-        $post = Post::find($id);
-
-        // Verifica se informou o arquivo e se é válido
-        if ($request->hasFile('image')) 
-        {
-           
-            if($request->image->isValid())
+        if(Auth::user()->type_user == 0) {
+            $request->validate([
+                'name' => ['string', 'max:100'],
+                'color' => ['string', 'max:100'],
+            ],[
+                'name.max' => 'Desculpe o nome excede o limite de caracteres.'
+            ]);
+            
+            $post = Post::find($id);
+    
+            // Verifica se informou o arquivo e se é válido
+            if ($request->hasFile('image')) 
             {
+               
+                if($request->image->isValid())
+                {
+                    $post->update(
+                        ['color' => $request->color, 'logo' => $request->logo, 'image' => $request->image->store('images')]
+                    );
+                }
+                    unset($image);
+            } else {
+                
                 $post->update(
-                    ['color' => $request->color, 'logo' => $request->logo, 'image' => $request->image->store('images')]
+                    ['color' => $request->color, 'logo' => $request->logo]
                 );
             }
-                unset($image);
+    
+            $campaign_post = CampaignPost::where('post_id', $id)->first();
+            $campaign_post->campaign_id = $request->campaign;
+            $campaign_post->save();
+    
+            return back()->with('success','Post atualizado com sucesso!');
         } else {
-            
-            $post->update(
-                ['color' => $request->color, 'logo' => $request->logo]
-            );
+            return back()->with('error','Usuário não autorizado.');
         }
-
-        $campaign_post = CampaignPost::where('post_id', $id)->first();
-        $campaign_post->campaign_id = $request->campaign;
-        $campaign_post->save();
-
-        return back()->with('success','Post atualizado com sucesso!');
     }
 
     /**
@@ -151,15 +160,19 @@ class PostController extends Controller
      */
     public function destroy($id) 
     {
-        $campaign_post = CampaignPost::where('post_id',$id)->first();
-        $campaign_post->delete();
-        
-        $post = Post::find($id);
-
-        Storage::disk('public')->delete($post->image);
-
-        $post->delete();
-
-        return redirect()->back()->with('success','Post excluído com sucesso!');
+        if(Auth::user()->type_user == 0) {
+            $campaign_post = CampaignPost::where('post_id',$id)->first();
+            $campaign_post->delete();
+            
+            $post = Post::find($id);
+    
+            Storage::disk('public')->delete($post->image);
+    
+            $post->delete();
+    
+            return redirect()->back()->with('success','Post excluído com sucesso!');
+        } else {
+            return back()->with('error','Usuário não autorizado.');
+        }
     }
 }
